Display node pool resource name and command correctly in CLI guide for inspecting node pools
