+++ src/components/MAPI/workernodes/guides/ListNodePoolsGuide.tsx
@@ -1,6 +1,7 @@
 import { Text } from 'grommet';
 import LoginGuideStep from 'MAPI/guides/LoginGuideStep';
 import { getCurrentInstallationContextName } from 'MAPI/guides/utils';
+import { Providers } from 'model/constants';
 import * as docs from 'model/constants/docs';
 import React from 'react';
 import CLIGuide from 'UI/Display/MAPI/CLIGuide';
@@ -8,22 +9,32 @@ import CLIGuideAdditionalInfo from 'UI/Display/MAPI/CLIGuide/CLIGuideAdditionalI
 import CLIGuideStep from 'UI/Display/MAPI/CLIGuide/CLIGuideStep';
 import CLIGuideStepList from 'UI/Display/MAPI/CLIGuide/CLIGuideStepList';
 
+function getProviderNodePoolResourceName(
+  provider: PropertiesOf<typeof Providers>
+) {
+  switch (provider) {
+    case Providers.AWS:
+    case Providers.GCP:
+      return 'MachineDeployment';
+    case Providers.AZURE:
+      return 'MachinePool';
+    default:
+      return 'MachinePool';
+  }
+}
 interface IListNodePoolsGuideProps
   extends Omit<React.ComponentPropsWithoutRef<typeof CLIGuide>, 'title'> {
   clusterName: string;
   clusterNamespace: string;
-  providerNodePoolResourceName: string;
+  provider: PropertiesOf<typeof Providers>;
 }
 
 const ListNodePoolsGuide: React.FC<
   React.PropsWithChildren<IListNodePoolsGuideProps>
-> = ({
-  clusterName,
-  clusterNamespace,
-  providerNodePoolResourceName,
-  ...props
-}) => {
+> = ({ clusterName, clusterNamespace, provider, ...props }) => {
   const context = getCurrentInstallationContextName();
+  const providerNodePoolResourceName =
+    getProviderNodePoolResourceName(provider);
 
   return (
     <CLIGuide
@@ -55,12 +66,19 @@ const ListNodePoolsGuide: React.FC<
         <LoginGuideStep />
         <CLIGuideStep
           title='2. List node pools'
-          command={`
-          kubectl gs --context ${context} \\
-            get nodepools \\
-            --cluster-name ${clusterName} \\
-            --namespace ${clusterNamespace}
-          `}
+          command={
+            provider === Providers.GCP
+              ? `
+              kubectl --context ${context} \\
+                get machinedeployments.cluster.x-k8s.io \\
+                --selector cluster.x-k8s.io/cluster-name=${clusterName},cluster.x-k8s.io/role!=bastion \\
+                --namespace ${clusterNamespace}`
+              : `
+              kubectl gs --context ${context} \\
+                get nodepools \\
+                --cluster-name ${clusterName} \\
+                --namespace ${clusterNamespace}`
+          }
         >
           <Text>
             <strong>Note:</strong> Add <code>--output json</code> to include
